Extract resolveLinkedAlbumId into @wxyc/database (BS#1829)
Moves resolveLinkedAlbumId, lookupKey, and flowsheetLookupKey out of apps/backend/services/album-metadata-lookup.service.ts into shared/database/src/album-resolve.ts so a jobs/ workspace (the upcoming album-critic-reviews-etl, #1830) can import the exact-match flowsheet resolver without reaching into apps/backend, which its Dockerfile build stage cannot see. Behavior-preserving move: the normalized (artist, album) lookup-key semantics and the deterministic most-recent-row-wins pick are unchanged. The service keeps a thin resolveLinkedAlbumId re-export so apps/backend/controllers/proxy.controller.ts and its test keep importing from the same path, mirroring the concerts-recompute.ts (BS#1763) and legacy-mirror (BS#1707) shim precedents. scripts/seed-critic-reviews.ts now imports the resolver from @wxyc/database directly.

The moved function's own test cases relocate from tests/unit/services/album-metadata-lookup.service.test.ts to a new tests/unit/database/album-resolve.test.ts, testing the real module directly (mocking only its client.js dependency, same shape as concerts-recompute.test.ts) rather than through the service's whole-package @wxyc/database mock, which can no longer exercise an implementation that now lives inside the mocked package. tests/mocks/database.mock.ts gains a resolveLinkedAlbumId stub matching the recomputeHasResolvedSupport precedent for any other consumer resolving the package through the default mock.
